Change function argument from short to int, to reflect implicit promotion.
git-svn-id: https://svn.code.sf.net/p/chicken/code/trunk/cotvnc@820 9ce1af06-8ae8-4ba5-be20-3b57e42882c5

+++ Source/d3des.c
@@ -71,7 +71,7 @@ static unsigned char pc2[48] = {
 
 void deskey(key, edf)	/* Thanks to James Gillogly & Phil Karn! */
 unsigned char *key;
-short edf;
+int edf;
 {
 	register int i, j, l, m, n;
 	unsigned char pc1m[56], pcr[56];
